Rolling back commit r3349 due to performace drop. Branch is now back to 3348.

+++ matrix/src/theory/arith/matrix.h
@@ -24,9 +24,8 @@
 #include "expr/node.h"
 #include "expr/attribute.h"
 
-#include "util/dense_map.h"
-
 #include "theory/arith/arithvar.h"
+#include "theory/arith/arithvar_set.h"
 #include "theory/arith/normal_form.h"
 
 #include <queue>
@@ -37,10 +36,10 @@ namespace CVC4 {
 namespace theory {
 namespace arith {
 
-typedef size_t EntryID;
+typedef uint32_t EntryID;
 const EntryID ENTRYID_SENTINEL = std::numeric_limits<EntryID>::max();
 
-typedef size_t RowIndex;
+typedef uint32_t RowIndex;
 const RowIndex ROW_INDEX_SENTINEL  = std::numeric_limits<RowIndex>::max();
 
 template<class T>
